Fix possible message corruption on long recv()s.
Moving the 'start' index of the ring buffer while scanning for message
delimiters marks the beginning of the message as free, meaning a long
recv() can overwrite it. Only move 'start' once we have the entire
message.

Triggered under Valgrind for the MOTD of some servers. Perhaps because
the slower speed allowed more data to accumulate.