Make 'serv_fd' a global variable.
Supporting multiple simultaneous IRC networks would be messy when
keeping track of state for e.g. reminders between invocations of the
bot. Assume instances of the bot always live on the same network.

Piggybacking fixes to the remind function. Now sends reminders either to
the channel or directly to a particular nick (in case the reminder was
set via a /msg to the bot).