Pass the "natural" reply target as a convenience.
For commands. The "natural" reply target is the channel for commands
written in a channel and the nick of the sender for commands sent
directly to the bot.

Move some parsing to before handle_cmd() to reduce code duplication.
Flag an empty argument as "missing" too (as there probably aren't that
many reasonable use cases for an empty argument). Get rid of
reply_target(). Make more stuff const.

Change how we detect direct messages to the bot to look for a channel as
the target. If the target does not look like a channel, assume it's a
direct message.

Get rid of UNUSED and turn off -Wunused-parameter as it seems to be more
trouble than it's worth.